requirements.txt
flask
flask-sqlalchemy
beautifulsoup4
requests
pytest